Refactored the category handling code into a custom hook
